pyproject.toml
[project]
name = "langgraph-memory"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "langchain>=1.2.13",
    "langchain-core>=1.2.20",
    "langchain-openai>=1.1.11",
    "langgraph>=1.1.3",
    "numpy>=2.4.3",
    "python-dotenv>=1.2.2",
]
